refactor: dedupe command and custom loader helpers
